refactor: e2e test for preset configurations
