+++ e2e-tests/authenticated/assetConfiguration.spec.ts
@@ -7,7 +7,6 @@ import { fromEnv } from '@nordicsemiconductor/from-env'
 import { expect, test } from '@playwright/test'
 import * as path from 'path'
 import { DataModules } from '../../src/asset/asset.js'
-import { presetConfigs } from '../../src/asset/config.js'
 import { checkForConsoleErrors } from '../lib/checkForConsoleErrors.js'
 import { loadSessionData } from '../lib/loadSessionData.js'
 import { AssetType, selectCurrentAsset } from './lib.js'
@@ -135,62 +134,4 @@ test("Should check 'update' button to be disabled when form is fill with null va
 	await expect(
 		page.locator('#asset-settings-form >> footer >> button'),
 	).not.toBeDisabled()
-})
-
-test("should charge preset values for 'Parcel' configuration", async ({
-	page,
-}) => {
-	await page.click('header[role="button"]:has-text("Settings")')
-
-	// select preset config
-	await page.locator('text=Parcel Config').click()
-
-	// TODO: check input fields to be updated
-
-	// update config
-	await page.click('#asset-settings-form >> footer >> button')
-
-	// Verify
-	const { thingName } = await loadSessionData(AssetType.Default)
-	const { payload } = await new IoTDataPlaneClient({
-		endpoint: `https://${mqttEndpoint}`,
-	}).send(
-		new GetThingShadowCommand({
-			thingName,
-		}),
-	)
-	expect(payload).not.toBeUndefined()
-	const shadow = JSON.parse(toUtf8(payload as Uint8Array))
-
-	// TODO: use presetConfigs as object to match
-	expect(shadow.state.desired.cfg).toMatchObject(presetConfigs.parcel.config)
-})
-
-test("should charge preset values for 'walking' configuration", async ({
-	page,
-}) => {
-	await page.click('header[role="button"]:has-text("Settings")')
-
-	// select preset config
-	await page.locator('text=Walking Config').click()
-
-	// TODO: check input fields to be updated
-
-	// update config
-	await page.click('#asset-settings-form >> footer >> button')
-
-	// Verify
-	const { thingName } = await loadSessionData(AssetType.Default)
-	const { payload } = await new IoTDataPlaneClient({
-		endpoint: `https://${mqttEndpoint}`,
-	}).send(
-		new GetThingShadowCommand({
-			thingName,
-		}),
-	)
-	expect(payload).not.toBeUndefined()
-	const shadow = JSON.parse(toUtf8(payload as Uint8Array))
-
-	// TODO: use presetConfigs as object to match
-	expect(shadow.state.desired.cfg).toMatchObject(presetConfigs.walking.config)
 })
